Pulled conversations into their own module to enforce separation from low-level messages
