+++ photonpump/__init__.py
@@ -1,3 +1,2 @@
-from .connection import *    # noqa
 from .exceptions import *    # noqa
 from .messages import *      # noqa
