+++ photonpump/messages.py
@@ -1,19 +1,14 @@
 import array
-import binascii
 import json
 import logging
 import math
 import struct
-from asyncio import Future, Queue, run_coroutine_threadsafe
 from collections import namedtuple
 from enum import IntEnum
-from typing import Any, Dict, NamedTuple, Sequence, Union
+from typing import Any, Dict, NamedTuple, Sequence
 from uuid import UUID, uuid4
 
-from google.protobuf import text_format
-from google.protobuf.message import DecodeError
-
-from . import exceptions, messages_pb2
+from . import messages_pb2
 
 HEADER_LENGTH = 1 + 1 + 16
 SIZE_UINT_32 = 4
@@ -97,22 +92,6 @@ SubscriptionDropReason = make_enum(
 )
 
 
-class ReplyAction(IntEnum):
-
-    CompleteScalar = 0
-    CompleteError = 1
-
-    BeginIterator = 2
-    YieldToIterator = 3
-    CompleteIterator = 4
-    RaiseToIterator = 5
-
-    BeginVolatileSubscription = 6
-    YieldToSubscription = 7
-    FinishSubscription = 8
-    RaiseToSubscription = 8
-
-
 class Credential:
 
     def __init__(self, username, password):
@@ -177,13 +156,6 @@ class OutboundMessage:
 
         self.length = HEADER_LENGTH
         self.is_authenticated = False
-
-
-class Reply(NamedTuple):
-
-    action: ReplyAction
-    result: Any
-    next_message: OutboundMessage
 
 
 class MessageReader:
@@ -284,89 +256,6 @@ class MessageReader:
                 )
 
 
-class Conversation:
-
-    def __init__(
-            self, conversation_id: UUID = None, credential: Credential = None
-    ) -> None:
-        self.conversation_id = conversation_id or uuid4()
-        self.result: Future = Future()
-        self.is_complete = False
-        self.credential = credential
-
-    def reply(self, response: InboundMessage) -> Reply:
-        pass
-
-    def error(self, exn: Exception):
-        return Reply(ReplyAction.CompleteError, exn, None)
-
-    def conversation_error(self, exn_type, response) -> Reply:
-        error = response.payload.decode('UTF-8')
-        exn = exn_type(self.conversation_id, error)
-
-        return self.error(exn)
-
-    def unhandled_message(self, response):
-        body = messages_pb2.NotHandled()
-        body.ParseFromString(response.payload)
-
-        if body.reason == NotHandledReason.NotReady:
-            exn = exceptions.NotReady(self.conversation_id)
-        elif body.reason == NotHandledReason.TooBusy:
-            exn = exceptions.TooBusy(self.conversation_id)
-        elif body.reason == NotHandledReason.NotMaster:
-            exn = exceptions.NotMaster(self.conversation_id)
-        else:
-            exn = exceptions.NotHandled(self.conversation_id, body.reason)
-
-        return self.error(exn)
-
-    def respond_to(self, response: InboundMessage) -> Reply:
-        try:
-            if response.command is TcpCommand.BadRequest:
-                return self.conversation_error(exceptions.BadRequest, response)
-            elif response.command is TcpCommand.NotAuthenticated:
-                return self.conversation_error(
-                    exceptions.NotAuthenticated, response
-                )
-            elif response.command is TcpCommand.NotHandled:
-                return self.unhandled_message(response)
-            else:
-                return self.reply(response)
-        except Exception as e:
-            return self.error(
-                exceptions.PayloadUnreadable(
-                    self.conversation_id, response.payload, e
-                )
-            )
-
-        return None
-
-    def cancel(self):
-        self.result.cancel()
-
-
-class HeartbeatConversation(Conversation):
-
-    def __init__(self, conversation_id: UUID) -> None:
-        super().__init__(conversation_id)
-
-    def start(self):
-
-        return OutboundMessage(
-            self.conversation_id, TcpCommand.HeartbeatResponse, b''
-        )
-
-
-class PingConversation(Conversation):
-
-    def start(self):
-        return OutboundMessage(self.conversation_id, TcpCommand.Ping, b'')
-
-    def reply(self, _: InboundMessage):
-        return Reply(ReplyAction.CompleteScalar, None, None)
-
-
 class ExpectedVersion(IntEnum):
     """Static values for concurrency control
 
@@ -433,50 +322,6 @@ def _json(self):
 
 
 EventRecord.json = _json
-
-
-class StreamingIterator:
-
-    def __init__(self, size):
-        self.items = Queue(maxsize=size)
-        self.finished = False
-        self.fut = None
-
-    async def __aiter__(self):
-        return self
-
-    async def enqueue_items(self, items):
-        for item in items:
-            await self.items.put(item)
-
-    async def enqueue(self, item):
-        await self.items.put(item)
-
-    async def __anext__(self):
-        if self.finished and self.items.empty():
-            raise StopAsyncIteration()
-        try:
-            next = await self.items.get()
-        except Exception as e:
-            raise StopAsyncIteration()
-
-        if isinstance(next, StopIteration):
-            raise StopAsyncIteration()
-
-        if isinstance(next, Exception):
-            raise next
-
-        return next
-
-    async def athrow(self, e):
-        await self.items.put(e)
-
-    async def asend(self, m):
-        await self.items.put(m)
-
-    def cancel(self):
-        self.finished = True
-        self.asend(StopIteration())
 
 
 class Event:
@@ -560,68 +405,6 @@ def _make_event(record: messages_pb2.ResolvedEvent):
     return Event(event, link)
 
 
-class Operation:
-    """The base class for requests to Eventstore.
-
-    Implementors have two responsibilities: they must serialize a byte-stream
-    request in the :meth:`~photonpump.messages.Operation.send` method, and
-    they must deserialize and handle the response in the
-    :meth:`~photonpump.messages.Operation.handle_response` method.
-    """
-
-    one_way = False
-
-    def __init__(self, credential: Credential = None):
-        self.credential = credential
-
-        if credential is None:
-            self.flags = OperationFlags.Empty
-        else:
-            self.flags = OperationFlags.Authenticated
-
-    def send(self, writer):
-        """
-        Write the byte-stream of this request to an instance of StreamWriter
-        """
-        header = self.make_header()
-        dump(header[4:], self.data)
-        writer.write(header)
-        writer.write(self.data)
-
-    def make_header(self):
-        """Build the byte-array representing the operation's header."""
-        buf = bytearray()
-        data_length = len(self.data)
-        authn_length = self.credential.length if self.credential is not None else 0
-        buf.extend(
-            struct.pack(
-                '<IBB', HEADER_LENGTH + authn_length + data_length,
-                self.command, self.flags
-            )
-        )
-        buf.extend(self.correlation_id.bytes_le)
-
-        if self.flags == OperationFlags.Authenticated:
-            buf.extend(self.credential.bytes)
-
-        return buf
-
-    async def handle_response(self, header, payload, writer):
-        """Handle the response from Eventstore.
-
-        Implementors can choose whether to return a single result,
-
-        return an async generator, or send a new Operation to the
-        :class:`photonpump.Connection`.
-        """
-        self.is_complete = True
-
-    def __repr__(self):
-        return "Operation %s (%s)" % (
-            self.__class__.__name__, self.correlation_id
-        )
-
-
 def NewEvent(
         type: str,
         id: UUID = uuid4(),
@@ -642,514 +425,11 @@ def NewEvent(
     return NewEventData(id, type, data, metadata)
 
 
-class WriteEventsConversation(Conversation):
-    """Command class for writing a sequence of events to a single
-        stream.
-
-    Args:
-        stream: The name of the stream to write to.
-        events: A sequence of events to write.
-        expected_version (optional): The expected version of the
-            target stream used for concurrency control.
-        required_master (optional): True if this command must be
-            sent direct to the master node, otherwise False.
-        correlation_id (optional): A unique identifer for this
-            command.
-
-    """
-
-    def __init__(
-            self,
-            stream: str,
-            events: Sequence[NewEventData],
-            expected_version: Union[ExpectedVersion, int] = ExpectedVersion.Any,
-            require_master: bool = False,
-            conversation_id: UUID = None,
-            credential=None,
-            loop=None
-    ):
-        super().__init__(conversation_id, credential)
-        self.stream = stream
-        self.require_master = require_master
-        self.events = events
-        self.expected_version = expected_version
-
-    def start(self):
-        msg = messages_pb2.WriteEvents()
-        msg.event_stream_id = self.stream
-        msg.require_master = self.require_master
-        msg.expected_version = self.expected_version
-
-        for event in self.events:
-            e = msg.events.add()
-            e.event_id = event.id.bytes_le
-            e.event_type = event.type
-
-            if isinstance(event.data, str):
-                e.data_content_type = ContentType.Json
-                e.data = event.data.encode('UTF-8')
-            elif event.data:
-                e.data_content_type = ContentType.Json
-                e.data = json.dumps(event.data).encode('UTF-8')
-            else:
-                e.data_content_type = ContentType.Binary
-                e.data = bytes()
-
-            if event.metadata:
-                e.metadata_content_type = ContentType.Json
-                e.metadata = json.dumps(event.metadata).encode('UTF-8')
-            else:
-                e.metadata_content_type = ContentType.Binary
-                e.metadata = bytes()
-
-        data = msg.SerializeToString()
-
-        return OutboundMessage(
-            self.conversation_id, TcpCommand.WriteEvents, data, self.credential
-        )
-
-    def reply(self, response: InboundMessage):
-        result = messages_pb2.WriteEventsCompleted()
-        result.ParseFromString(response.payload)
-        self.is_complete = True
-        self.result.set_result(result)
-
-
 ReadEventResult = make_enum(messages_pb2._READEVENTCOMPLETED_READEVENTRESULT)
 
 ReadStreamResult = make_enum(
     messages_pb2._READSTREAMEVENTSCOMPLETED_READSTREAMRESULT
 )
-
-
-class ReadStreamEventsBehaviour:
-
-    def __init__(self, result_type, response_cls):
-        self.result_type = result_type
-        self.response_cls = response_cls
-
-    def success(self, result):
-        pass
-
-    def error(self, exn: Exception):
-        pass
-
-    def reply(self, response: InboundMessage):
-        result = self.response_cls()
-        result.ParseFromString(response.payload)
-
-        if result.result == self.result_type.Success:
-            return self.success(result)
-        elif result.result == self.result_type.NoStream:
-            return self.error(
-                exceptions.StreamNotFound(self.conversation_id, self.stream)
-            )
-        elif result.result == self.result_type.StreamDeleted:
-            return self.error(
-                exceptions.StreamDeleted(self.conversation_id, self.stream)
-            )
-        elif result.result == self.result_type.Error:
-            return self.error(
-                exceptions.ReadError(
-                    self.conversation_id, self.stream, result.error
-                )
-            )
-        elif result.result == self.result_type.AccessDenied:
-            return self.error(
-                exceptions.AccessDenied(
-                    self.conversation_id,
-                    type(self).__name__,
-                    result.error,
-                    stream=self.stream
-                )
-            )
-        elif self.result_type == ReadEventResult and result.result == self.result_type.NotFound:
-            return self.error(
-                exceptions.EventNotFound(
-                    self.conversation_id, self.stream, self.event_number
-                )
-            )
-
-
-class ReadEventConversation(ReadStreamEventsBehaviour, Conversation):
-    """Command class for reading a single event.
-
-    Args:
-        stream: The name of the stream containing the event.
-        event_number: The sequence number of the event to read.
-        resolve_links (optional): True if eventstore should
-            automatically resolve Link Events, otherwise False.
-        required_master (optional): True if this command must be
-            sent direct to the master node, otherwise False.
-        correlation_id (optional): A unique identifer for this
-            command.
-
-    """
-
-    def __init__(
-            self,
-            stream: str,
-            event_number: int,
-            resolve_links: bool = True,
-            require_master: bool = False,
-            conversation_id: UUID = None,
-            credentials=None
-    ) -> None:
-
-        Conversation.__init__(self, conversation_id, credential=credentials)
-        ReadStreamEventsBehaviour.__init__(
-            self, ReadEventResult, messages_pb2.ReadEventCompleted
-        )
-        self.stream = stream
-        self.event_number = event_number
-        self.require_master = require_master
-        self.resolve_link_tos = resolve_links
-
-    def start(self) -> OutboundMessage:
-        msg = messages_pb2.ReadEvent()
-        msg.event_number = self.event_number
-        msg.event_stream_id = self.stream
-        msg.require_master = self.require_master
-        msg.resolve_link_tos = self.resolve_link_tos
-
-        data = msg.SerializeToString()
-
-        return OutboundMessage(
-            self.conversation_id, TcpCommand.Read, data, self.credential
-        )
-
-    def success(self, response):
-        return Reply(
-            ReplyAction.CompleteScalar, _make_event(response.event), None
-        )
-
-    def error(self, exn):
-        return Reply(ReplyAction.CompleteError, exn, None)
-
-
-class ReadStreamEventsConversation(ReadStreamEventsBehaviour, Conversation):
-    """Command class for reading events from a stream.
-
-    Args:
-        stream: The name of the stream containing the event.
-        event_number: The sequence number of the event to read.
-        resolve_links (optional): True if eventstore should
-            automatically resolve Link Events, otherwise False.
-        required_master (optional): True if this command must be
-            sent direct to the master node, otherwise False.
-        correlation_id (optional): A unique identifer for this
-            command.
-    """
-
-    def __init__(
-            self,
-            stream: str,
-            from_event: int = 0,
-            max_count: int = 100,
-            resolve_links: bool = True,
-            require_master: bool = False,
-            direction: StreamDirection = StreamDirection.Forward,
-            credentials=None,
-            conversation_id: UUID = None
-    ) -> None:
-
-        Conversation.__init__(self, conversation_id, credential=credentials)
-        ReadStreamEventsBehaviour.__init__(
-            self, ReadStreamResult, messages_pb2.ReadStreamEventsCompleted
-        )
-        self.stream = stream
-        self.direction = direction
-        self.from_event = from_event
-        self.max_count = max_count
-        self.require_master = require_master
-        self.resolve_link_tos = resolve_links
-
-    def _fetch_page_message(self, from_event):
-        if self.direction == StreamDirection.Forward:
-            command = TcpCommand.ReadStreamEventsForward
-        else:
-            command = TcpCommand.ReadStreamEventsBackward
-
-        msg = messages_pb2.ReadStreamEvents()
-        msg.event_stream_id = self.stream
-        msg.from_event_number = from_event
-        msg.max_count = self.max_count
-        msg.require_master = self.require_master
-        msg.resolve_link_tos = self.resolve_link_tos
-
-        data = msg.SerializeToString()
-
-        return OutboundMessage(
-            self.conversation_id, command, data, self.credential
-        )
-
-    def start(self):
-        return self._fetch_page_message(self.from_event)
-
-    def success(self, result: messages_pb2.ReadStreamEventsCompleted):
-        events = [_make_event(x) for x in result.events]
-
-        return Reply(
-            ReplyAction.CompleteScalar,
-            StreamSlice(
-                events, result.next_event_number, result.last_event_number,
-                None, result.last_commit_position, result.is_end_of_stream
-            ), None
-        )
-
-    def error(self, exn):
-        return Reply(ReplyAction.CompleteError, exn, None)
-
-
-class IterStreamEvents(ReadStreamEventsBehaviour, Conversation):
-    """Command class for iterating events from a stream.
-
-    Args:
-        stream: The name of the stream containing the event.
-        resolve_links (optional): True if eventstore should
-            automatically resolve Link Events, otherwise False.
-        required_master (optional): True if this command must be
-            sent direct to the master node, otherwise False.
-        correlation_id (optional): A unique identifer for this
-            command.
-
-    """
-
-    def __init__(
-            self,
-            stream: str,
-            from_event: int = 0,
-            batch_size: int = 100,
-            resolve_links: bool = True,
-            require_master: bool = False,
-            direction: StreamDirection = StreamDirection.Forward,
-            credentials=None,
-            conversation_id: UUID = None
-    ):
-
-        Conversation.__init__(self, conversation_id, credentials)
-        ReadStreamEventsBehaviour.__init__(
-            self, ReadStreamResult, messages_pb2.ReadStreamEventsCompleted
-        )
-        self.batch_size = batch_size
-        self.waiting_for_page = 0
-        self.stream = stream
-        self.resolve_links = resolve_links
-        self.require_master = require_master
-        self.direction = direction
-        self.from_event = from_event
-
-        if direction == StreamDirection.Forward:
-            self.command = TcpCommand.ReadStreamEventsForward
-        else:
-            self.command = TcpCommand.ReadStreamEventsBackward
-
-    def _fetch_page_message(self, from_event):
-        if self.direction == StreamDirection.Forward:
-            command = TcpCommand.ReadStreamEventsForward
-        else:
-            command = TcpCommand.ReadStreamEventsBackward
-
-        msg = messages_pb2.ReadStreamEvents()
-        msg.event_stream_id = self.stream
-        msg.from_event_number = from_event
-        msg.max_count = self.batch_size
-        msg.require_master = self.require_master
-        msg.resolve_link_tos = self.resolve_links
-
-        data = msg.SerializeToString()
-
-        return OutboundMessage(
-            self.conversation_id, command, data, self.credential
-        )
-
-    def start(self):
-        return self._fetch_page_message(self.from_event)
-
-    def success(self, result: messages_pb2.ReadStreamEventsCompleted):
-        events = [_make_event(x) for x in result.events]
-
-        next_message = self._fetch_page_message(
-            result.next_event_number
-        ) if not result.is_end_of_stream else None
-
-        self.waiting_for_page += 1
-
-        return Reply(
-            ReplyAction.BeginIterator,
-            StreamSlice(
-                events, result.next_event_number, result.last_event_number,
-                None, result.last_commit_position, result.is_end_of_stream
-            ), next_message
-        )
-
-    def error(self, exn: Exception) -> Reply:
-        if self.waiting_for_page > 0:
-            return Reply(ReplyAction.RaiseToIterator, exn, None)
-
-        return Reply(ReplyAction.CompleteError, exn, None)
-
-
-class VolatileSubscription:
-
-    def __init__(
-            self, stream, initial_commit, initial_event_number, buffer_size
-    ):
-        self.last_commit_position = initial_commit
-        self.last_event_number = initial_event_number
-        self.events = StreamingIterator(buffer_size)
-        self.stream = stream
-
-    async def enqueue(self, commit_position, event):
-        self.last_commit_position = commit_position
-        self.last_event_number = event.original_event.event_number
-        await self.events.enqueue(event)
-
-    def cancel(self):
-        self.events.cancel()
-
-
-class PersistentSubscription:
-
-    def __init__(
-            self,
-            conn,
-            name,
-            stream,
-            correlation_id,
-            initial_commit,
-            initial_event_number,
-            buffer_size,
-            auto_ack=False
-    ):
-        self.initial_commit_position = initial_commit
-        self.name = name
-        self.correlation_id = correlation_id
-        self.last_event_number = initial_event_number
-        self.events = StreamingIterator(buffer_size)
-        self.stream = stream
-        self.auto_ack = auto_ack
-        self.conn = conn
-
-    async def enqueue(self, event):
-        self.last_event_number = event.original_event.event_number
-        await self.events.enqueue(event)
-
-        if self.auto_ack:
-            await self.conn.ack(event)
-
-    async def ack(self, message: Event):
-        await self.conn.ack(
-            self.name,
-            message.original_event_id,
-            correlation_id=self.correlation_id
-        )
-
-    def cancel(self):
-        self.events.cancel()
-
-
-class CreateVolatileSubscription(Conversation):
-    """Command class for creating a non-persistent subscription.
-
-    Args:
-        stream: The name of the stream to watch for new events
-    """
-
-    class State(IntEnum):
-        init = 0
-        catch_up = 1
-        live = 2
-
-    def __init__(
-            self,
-            stream: str,
-            resolve_links: bool = True,
-            buffer_size: int = 1,
-            credentials: Credential = None,
-            conversation_id: UUID = None
-    ) -> None:
-        super().__init__(conversation_id, credentials)
-        self.stream = stream
-        self.buffer_size = buffer_size
-        self.resolve_links = resolve_links
-        self.state = CreateVolatileSubscription.State.init
-
-    def error(self, exn):
-        return Reply(ReplyAction.RaiseToSubscription, exn, None)
-
-    def start(self):
-        msg = messages_pb2.SubscribeToStream()
-        msg.event_stream_id = self.stream
-        msg.resolve_link_tos = self.resolve_links
-
-        return OutboundMessage(
-            self.conversation_id, TcpCommand.SubscribeToStream,
-            msg.SerializeToString()
-        )
-
-    def reply_from_init(self, response: InboundMessage):
-        result = messages_pb2.SubscriptionConfirmation()
-        result.ParseFromString(response.payload)
-
-        self.state = CreateVolatileSubscription.State.live
-
-        return Reply(
-            ReplyAction.BeginVolatileSubscription,
-            VolatileSubscription(
-                self.stream, result.last_commit_position,
-                result.last_event_number, self.buffer_size
-            ), None
-        )
-
-    def reply_from_live(self, response: InboundMessage):
-        result = messages_pb2.StreamEventAppeared()
-        result.ParseFromString(response.payload)
-
-        return Reply(
-            ReplyAction.YieldToSubscription, _make_event(result.event), None
-        )
-
-    def reply(self, response: InboundMessage):
-
-        if self.state == CreateVolatileSubscription.State.init:
-            return self.reply_from_init(response)
-
-        if self.state == CreateVolatileSubscription.State.live:
-            return self.reply_from_live(response)
-
-    async def handle_response(self, header, payload, writer):
-        if header.cmd == TcpCommand.SubscriptionConfirmation.value:
-            result = messages_pb2.SubscriptionConfirmation()
-            result.ParseFromString(payload)
-            self.subscription = VolatileSubscription(
-                self.stream, result.last_commit_position,
-                result.last_event_number, self.buffer_size
-            )
-
-            self.future.set_result(self.subscription)
-
-        elif header.cmd == TcpCommand.StreamEventAppeared:
-            result = messages_pb2.StreamEventAppeared()
-            try:
-                result.ParseFromString(payload)
-                event = _make_event(result.event)
-                await self.subscription.enqueue(
-                    event.original_event.commit_position, event
-                )
-            except Exception as e:
-                logging.debug(e)
-                logging.debug(payload)
-                logging.debug(header)
-
-        elif header.cmd == TcpCommand.SubscriptionDropped:
-            self.subscription.cancel()
-
-    def cancel(self):
-        if not self.future.done():
-            self.future.cancel()
-        else:
-            self.subscription.cancel()
 
 
 SubscriptionResult = make_enum(
@@ -1161,140 +441,3 @@ SubscriptionResult = make_enum(
 class SubscriptionCreatedResponse(NamedTuple):
     result: SubscriptionResult
     reason: str
-
-
-class CreatePersistentSubscription(Operation):
-
-    def __init__(
-            self,
-            name,
-            stream,
-            resolve_links=True,
-            start_from=-1,
-            timeout_ms=8192,
-            record_statistics=False,
-            live_buffer_size=128,
-            read_batch_size=128,
-            buffer_size=128,
-            max_retry_count=3,
-            prefer_round_robin=True,
-            checkpoint_after_ms=1024,
-            checkpoint_max_count=1024,
-            checkpoint_min_count=10,
-            subscriber_max_count=10,
-            credentials=None,
-            correlation_id=None,
-            loop=None
-    ) -> None:
-        self.stream = stream
-        self.name = name
-        msg = messages_pb2.CreatePersistentSubscription()
-        msg.subscription_group_name = name
-        msg.event_stream_id = stream
-        msg.start_from = start_from
-        msg.resolve_link_tos = resolve_links
-        msg.message_timeout_milliseconds = timeout_ms
-        msg.record_statistics = record_statistics
-        msg.live_buffer_size = live_buffer_size
-        msg.read_batch_size = read_batch_size
-        msg.buffer_size = buffer_size
-        msg.max_retry_count = max_retry_count
-        msg.prefer_round_robin = prefer_round_robin
-        msg.checkpoint_after_time = checkpoint_after_ms
-        msg.checkpoint_max_count = checkpoint_max_count
-        msg.checkpoint_min_count = checkpoint_min_count
-        msg.subscriber_max_count = subscriber_max_count
-
-        self.command = TcpCommand.CreatePersistentSubscription
-        self.future: Future = Future(loop=loop)
-        self.data = msg.SerializeToString()
-        self.correlation_id = correlation_id or uuid4()
-        self.is_complete = False
-        super().__init__(credentials)
-
-    async def handle_response(self, header, payload, writer):
-
-        result = messages_pb2.CreatePersistentSubscriptionCompleted()
-        result.ParseFromString(payload)
-        self.future.set_result(
-            SubscriptionCreatedResponse(result.result, result.reason)
-        )
-
-    def cancel(self):
-        self.future.cancel()
-
-
-class ConnectPersistentSubscription(Operation):
-
-    def __init__(
-            self,
-            name,
-            stream,
-            connection,
-            max_in_flight=10,
-            credentials=None,
-            correlation_id=None,
-            loop=None
-    ) -> None:
-        self.stream = stream
-        self.max_in_flight = max_in_flight
-        self.name = name
-        self.conn = connection
-        msg = messages_pb2.ConnectToPersistentSubscription()
-        msg.subscription_id = name
-        msg.event_stream_id = stream
-        msg.allowed_in_flight_messages = max_in_flight
-
-        self.command = TcpCommand.ConnectToPersistentSubscription
-        self.future: Future = Future(loop=loop)
-        self.data = msg.SerializeToString()
-        self.correlation_id = correlation_id or uuid4()
-        self.is_complete = False
-        super().__init__(credentials)
-
-    def createSubscription(self, payload):
-        result = messages_pb2.PersistentSubscriptionConfirmation()
-        result.ParseFromString(payload)
-        self.subscription = PersistentSubscription(
-            self.conn, result.subscription_id, self.stream, self.correlation_id,
-            result.last_commit_position, result.last_event_number,
-            self.max_in_flight
-        )
-
-        self.future.set_result(self.subscription)
-
-    async def yieldEvent(self, payload):
-        result = messages_pb2.PersistentSubscriptionStreamEventAppeared()
-        result.ParseFromString(payload)
-        try:
-            await self.subscription.enqueue(_make_event(result.event))
-        except Exception as e:
-            print(e)
-
-    async def handle_response(self, header, payload, writer):
-        try:
-            if header.cmd == TcpCommand.PersistentSubscriptionConfirmation:
-                self.createSubscription(payload)
-            elif header.cmd == TcpCommand.PersistentSubscriptionStreamEventAppeared:
-                await self.yieldEvent(payload)
-        except Exception as e:
-            print(e)
-
-    def cancel(self):
-        self.future.cancel()
-
-
-class AcknowledgeMessages(Operation):
-
-    one_way = True
-
-    def __init__(
-            self, name, message_ids, correlation_id, credentials=None, loop=None
-    ) -> None:
-        msg = messages_pb2.PersistentSubscriptionAckEvents()
-        msg.subscription_id = name
-        msg.processed_event_ids.extend([id.bytes_le for id in message_ids])
-        self.command = TcpCommand.PersistentSubscriptionAckEvents
-        self.data = msg.SerializeToString()
-        self.correlation_id = correlation_id or uuid4()
-        super().__init__(credentials)
