+++ test/conversations/test_error_handling.py
@@ -3,6 +3,7 @@ from uuid import uuid4
 import photonpump.exceptions as exn
 import photonpump.messages as msg
 import photonpump.messages_pb2 as proto
+from photonpump.conversations import Ping, ReplyAction,WriteEventsConversation
 
 
 def test_bad_request():
@@ -15,7 +16,7 @@ def test_bad_request():
     data = {'pony_name': 'Burning Sulphur', 'distance': 6}
     event_data = msg.NewEventData(event_id, event_type, data, None)
 
-    conversation = msg.WriteEventsConversation(
+    conversation = WriteEventsConversation(
         "my-stream", [event_data], conversation_id=conversation_id
     )
 
@@ -27,7 +28,7 @@ def test_bad_request():
         )
     )
 
-    assert result.action == msg.ReplyAction.CompleteError
+    assert result.action == ReplyAction.CompleteError
     assert isinstance(result.result, exn.BadRequest)
     assert (result.result.message == error_message)
 
@@ -42,7 +43,7 @@ def test_not_authenticated():
     data = {'pony_name': 'Burning Sulphur', 'distance': 6}
     event_data = msg.NewEventData(event_id, event_type, data, None)
 
-    conversation = msg.WriteEventsConversation(
+    conversation = WriteEventsConversation(
         "my-stream", [event_data], conversation_id=conversation_id
     )
 
@@ -54,7 +55,7 @@ def test_not_authenticated():
         )
     )
 
-    assert reply.action == msg.ReplyAction.CompleteError
+    assert reply.action == ReplyAction.CompleteError
     assert isinstance(reply.result, exn.NotAuthenticated)
     assert reply.result.message == error_message
 
@@ -63,13 +64,13 @@ def test_notready_message():
 
     payload = proto.NotHandled()
     payload.reason = msg.NotHandledReason.NotReady
-    conversation = msg.PingConversation()
+    conversation = Ping()
     reply = conversation.respond_to(
         msg.InboundMessage(
             uuid4(), msg.TcpCommand.NotHandled, payload.SerializeToString()))
 
     error = reply.result
-    assert reply.action == msg.ReplyAction.CompleteError
+    assert reply.action == ReplyAction.CompleteError
     assert isinstance(error, exn.NotReady)
     assert error.conversation_id == conversation.conversation_id
 
@@ -78,12 +79,12 @@ def test_too_busy_message():
 
     payload = proto.NotHandled()
     payload.reason = msg.NotHandledReason.TooBusy
-    conversation = msg.PingConversation()
+    conversation = Ping()
     reply = conversation.respond_to(
         msg.InboundMessage(
             uuid4(), msg.TcpCommand.NotHandled, payload.SerializeToString()))
 
-    assert reply.action == msg.ReplyAction.CompleteError
+    assert reply.action == ReplyAction.CompleteError
     error = reply.result
     assert isinstance(error, exn.TooBusy)
     assert error.conversation_id == conversation.conversation_id
@@ -93,12 +94,12 @@ def test_not_master():
 
     payload = proto.NotHandled()
     payload.reason = msg.NotHandledReason.NotMaster
-    conversation = msg.PingConversation()
+    conversation = Ping()
     reply = conversation.respond_to(
         msg.InboundMessage(
             uuid4(), msg.TcpCommand.NotHandled, payload.SerializeToString()))
 
-    assert reply.action == msg.ReplyAction.CompleteError
+    assert reply.action == ReplyAction.CompleteError
     error = reply.result
     assert isinstance(error, exn.NotMaster)
     assert error.conversation_id == conversation.conversation_id
@@ -106,12 +107,12 @@ def test_not_master():
 
 def test_decode_error():
 
-    conversation = msg.PingConversation()
+    conversation = Ping()
     reply = conversation.respond_to(
         msg.InboundMessage(
             uuid4(), msg.TcpCommand.NotHandled, b'\x08\2A'))
 
-    assert reply.action == msg.ReplyAction.CompleteError
+    assert reply.action == ReplyAction.CompleteError
     error = reply.result
     assert isinstance(error, exn.PayloadUnreadable)
     assert error.conversation_id == conversation.conversation_id
