+++ test/conversations/test_write_events_conversation.py
@@ -4,6 +4,7 @@ from uuid import uuid4
 import photonpump.exceptions as exn
 import photonpump.messages as msg
 import photonpump.messages_pb2 as proto
+from photonpump.conversations import WriteEventsConversation
 
 
 def test_write_one_event():
@@ -15,7 +16,7 @@ def test_write_one_event():
     data = {'pony_name': 'Burning Sulphur', 'distance': 6}
     event_data = msg.NewEventData(event_id, event_type, data, None)
 
-    conversation = msg.WriteEventsConversation(
+    conversation = WriteEventsConversation(
         "my-stream", [event_data], conversation_id=conversation_id
     )
 
@@ -54,7 +55,7 @@ def test_one_event_response():
     data = {'pony_name': 'Burning Sulphur', 'distance': 6}
     event_data = msg.NewEventData(event_id, event_type, data, None)
 
-    conversation = msg.WriteEventsConversation(
+    conversation = WriteEventsConversation(
         "my-stream", [event_data], conversation_id=conversation_id
     )
